Simple Calculator with Match Case

+++ control-flow/match_case_calculator.py
@@ -1,5 +1,5 @@
-num1 = float(input("Enter first number: "))
-num2 = float(input("Enter second number: "))
+num1 = int(input("Enter first number: "))
+num2 = int(input("Enter second number: "))
 operation = input("Choose the operation (+, -, /, *): ")
 
 result = num1 + num2
